% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{cgrdpg-package}
\alias{cgrdpg}
\alias{cgrdpg-package}
\title{cgrdpg: GRDPG with contextual covariates}
\description{
Tools to estimate latent positions X and covariate loadings Z in a generalized
random dot product graph model with Gaussian covariates B = Z X^T + W,
using adjacency spectral embedding and Fisher scoring. This package is meant
as a research prototype accompanying a manuscript.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://example.com}
  \item Report bugs at \url{https://example.com/issues}
}

}
\author{
\strong{Maintainer}: Myung-Jin Lee \email{your.email@example.com}

}
\keyword{internal}
